all_for_one/indianExpress.py
import json
import os
import time
import mysql.connector
from selenium import webdriver
from selenium.webdriver.common.by import By
from selenium.webdriver.support.ui import WebDriverWait
from selenium.webdriver.support import expected_conditions as EC

def scrape_indian_express(driver):
    driver.get("https://indianexpress.com/")
    time.sleep(5)
    list = []
    news_list = []
    
    # Your scraping logic...
    
    return news_list

def save_news_to_file(driver, file_path="indian_express_all_news.json"):
    # Your logic to save news...
    
def store_news_in_db(news_list):
    # Your logic to store in DB...

def main():
    driver = webdriver.Firefox()
    try:
        save_news_to_file(driver)
    finally:
        driver.quit()

if __name__ == "__main__":
    main()