all_for_one/requirements.txt
Flask
flask-mysqldb
transformers
torch
selenium
beautifulsoup4
pandas
matplotlib
seaborn
mysql-connector-python